orchestrator-node/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
aiosqlite==0.20.0
numpy==2.1.2
python-dotenv==1.0.1
PyYAML==6.0.2
pytest==8.3.3
pytest-asyncio==0.24.0
